% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{n_expressoExpr}
\alias{n_expressoExpr}
\title{Micoarray expression data used in NeuroExpresso.}
\format{An object of class \code{data.frame} with 11593 rows and 369 columns.}
\usage{
n_expressoExpr
}
\description{
Micoarray expression data used in NeuroExpresso.
}
\keyword{datasets}
