% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{n_expressoSamples}
\alias{n_expressoSamples}
\title{Metadata for microarray expression data used in NeuroExpresso}
\format{An object of class \code{data.frame} with 363 rows and 21 columns.}
\usage{
n_expressoSamples
}
\description{
Metadata for microarray expression data used in NeuroExpresso
}
\keyword{datasets}
