% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{regionHierarchy}
\alias{regionHierarchy}
\title{Hierarchy used to classify brain regions}
\format{An object of class \code{list} of length 1.}
\usage{
regionHierarchy
}
\description{
Hierarchy used to classify brain regions
}
\keyword{datasets}
