% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{TasicMouseExp}
\alias{TasicMouseExp}
\title{Expression data from Tasic et al. 2016 single cell data}
\format{An object of class \code{matrix} with 22188 rows and 1809 columns.}
\usage{
TasicMouseExp
}
\description{
Expression data from Tasic et al. 2016 single cell data
}
\keyword{datasets}
