% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{TasicMouseMeta}
\alias{TasicMouseMeta}
\title{Metadata for Tasic et al. 2016 single cell data}
\format{An object of class \code{data.frame} with 1809 rows and 13 columns.}
\usage{
TasicMouseMeta
}
\description{
Metadata for Tasic et al. 2016 single cell data
}
\keyword{datasets}
